requirements.txt
discord.py
python-dotenv
openai
asyncio
logging
asyncpg
